Move relevant files from lib/ to 'app/lib/' (#7678)
* file renames

* point path to new file location

* Move rake tasks back to lib/tasks

* Move proficiency_cutoffs.json to app/lib/data

* Appease rubocop

* Fix javascript import path

* Move eg_form_builder back to lib

* Move app/lib/data back to lib/data

* Move AsyncHelper to spec/support
