+++ services/QuillLMS/app/lib/clever_library/api/oauth.rb
