+++ services/QuillLMS/app/lib/feature.rb
