+++ services/QuillLMS/app/lib/name_generator.rb
