+++ services/QuillLMS/spec/workers/fast_assign_worker_spec.rb
@@ -33,6 +33,4 @@ describe FastAssignWorker, type: :worker do
       expect(unit.classroom_units.find(original_classroom_unit.id).assign_on_join).to eq(true)
     end
   end
-
-
 end
